pyproject.toml
[project]
name = "jobify"
version = "0.1.0"
description = "Jobify — AI-Powered Interview Coach"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.9.1",
    "certifi==2025.7.9",
    "charset-normalizer==3.4.2",
    "django==5.2.4",
    "djangorestframework==3.16.0",
    "gunicorn==23.0.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "packaging==25.0",
    "pluggy==1.6.0",
    "psycopg2-binary==2.9.10",
    "pygments==2.19.2",
    "pytest==8.4.1",
    "pytest-django==4.11.1",
    "python-dotenv==1.1.1",
    "requests==2.32.4",
    "sqlparse==0.5.3",
    "urllib3==2.5.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "jobify_backend.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["backend"]
addopts = [
    "--tb=no",
    "--strict-markers",
]
